requirements.txt
alive_progress==3.1.4
catkin_pkg==1.0.0
cv_bridge==1.13.0.post0
einops==0.7.0
icecream==2.1.3
imageio==2.28.0
kornia==0.7.0
matplotlib==3.7.1
numpy==1.24.3
nvidia_smi==0.1.3
open3d_python==0.3.0.0
opencv_python==4.7.0.72
pandas==2.0.1
progress==1.6
pypcd4==0.4.5
pyserial==3.5
robotathome==1.1.9
scipy==1.12.0
taichi_nightly==1.7.0.post20230921
torch==2.0.1
torchmetrics==1.1.2
tqdm==4.65.0
